Modified the makefiles to use the log files - *.log - as the indicator of whether a text file has been loaded. We no longer use a zero-length *.?ld files to indicate this state. And we just run sqlplus to delete the old text from the database and don't worry about whether it is actually loaded already. If it isn't, sqlplus will just delete 0 rows.
We also added a new target to the makefiles: up-to-date. This causes
all *.log files to be touched, so that it appears to subsequent makes
that all text have have been loaded. This might be useful is a new clone
of the repo has been created, since there will be no *.log files in it.